Move NvtFamilies to an own js module
Extract nvt families components from the edit scan config dialog module
into an own one.
