+++ gsa/src/web/pages/scanconfigs/editdialog.js
@@ -24,233 +24,23 @@ import _ from 'gmp/locale';
 import {
   OPENVAS_SCAN_CONFIG_TYPE,
   OSP_SCAN_CONFIG_TYPE,
-  SCANCONFIG_TREND_DYNAMIC,
-  SCANCONFIG_TREND_STATIC,
-  parseTrend,
 } from 'gmp/models/scanconfig';
-
-import {isDefined} from 'gmp/utils/identity';
-
-import {YES_VALUE, NO_VALUE} from 'gmp/parser';
 
 import PropTypes from 'web/utils/proptypes';
 import {renderSelectItems} from 'web/utils/render';
 
 import SaveDialog from 'web/components/dialog/savedialog';
 
-import Checkbox from 'web/components/form/checkbox';
 import FormGroup from 'web/components/form/formgroup';
-import Radio from 'web/components/form/radio';
 import TextField from 'web/components/form/textfield';
 import Select from 'web/components/form/select';
 import Loading from 'web/components/loading/loading';
 
-import EditIcon from 'web/components/icon/editicon';
-
-import Divider from 'web/components/layout/divider';
 import Layout from 'web/components/layout/layout';
 
-import Section from 'web/components/section/section';
-
-import Table from 'web/components/table/stripedtable';
-import TableBody from 'web/components/table/body';
-import TableData from 'web/components/table/data';
-import TableHeader from 'web/components/table/header';
-import TableHead from 'web/components/table/head';
-import TableRow from 'web/components/table/row';
-
+import NvtFamilies from './nvtfamilies';
 import NvtPreferences from './nvtpreferences';
 import ScannerPreferences from './scannerpreferences';
-import Trend from './trend';
-
-class NvtFamily extends React.Component {
-  shouldComponentUpdate(nextProps, nextState) {
-    return (
-      nextProps.config !== this.props.config ||
-      nextProps.family !== this.props.family ||
-      nextProps.select !== this.props.select ||
-      nextProps.trend !== this.props.trend
-    );
-  }
-
-  render() {
-    const {
-      config,
-      family,
-      select,
-      trend,
-      onEditConfigFamilyClick,
-      onSelectChange,
-      onTrendChange,
-    } = this.props;
-    const {name} = family;
-    const config_family = config.families[name];
-    const counts = {
-      count: 0,
-      max: family.maxNvtCount,
-    };
-
-    if (isDefined(config_family)) {
-      counts.count = config_family.nvts.count;
-      counts.max = config_family.nvts.max;
-    }
-
-    const title =
-      config.usage_type === 'policy'
-        ? _('Edit Policy Family')
-        : _('Edit Scan Config Family');
-
-    return (
-      <TableRow key={name}>
-        <TableData>{name}</TableData>
-        <TableData align="start">{_('{{count}} of {{max}}', counts)}</TableData>
-        <TableData align={['center', 'start']}>
-          <Divider>
-            <Radio
-              flex
-              name={name}
-              checked={trend === SCANCONFIG_TREND_DYNAMIC}
-              convert={parseTrend}
-              value={SCANCONFIG_TREND_DYNAMIC}
-              onChange={onTrendChange}
-            />
-            <Trend trend={SCANCONFIG_TREND_DYNAMIC} />
-            <Radio
-              flex
-              name={name}
-              checked={trend === SCANCONFIG_TREND_STATIC}
-              convert={parseTrend}
-              value={SCANCONFIG_TREND_STATIC}
-              onChange={onTrendChange}
-            />
-            <Trend trend={SCANCONFIG_TREND_STATIC} />
-          </Divider>
-        </TableData>
-        <TableData align={['start', 'center']}>
-          <Checkbox
-            flex
-            name={name}
-            checked={select === YES_VALUE}
-            checkedValue={YES_VALUE}
-            unCheckedValue={NO_VALUE}
-            onChange={onSelectChange}
-          />
-        </TableData>
-        <TableData align={['center', 'center']}>
-          <EditIcon
-            title={title}
-            value={{familyName: name, configId: config.id}}
-            onClick={onEditConfigFamilyClick}
-          />
-        </TableData>
-      </TableRow>
-    );
-  }
-}
-
-NvtFamily.propTypes = {
-  config: PropTypes.model.isRequired,
-  family: PropTypes.object.isRequired,
-  select: PropTypes.yesno.isRequired,
-  trend: PropTypes.yesno.isRequired,
-  onEditConfigFamilyClick: PropTypes.func,
-  onSelectChange: PropTypes.func,
-  onTrendChange: PropTypes.func,
-};
-
-class NvtFamilies extends React.Component {
-  constructor(...args) {
-    super(...args);
-
-    this.onSelectChange = this.onSelectChange.bind(this);
-    this.onTrendChange = this.onTrendChange.bind(this);
-  }
-
-  onSelectChange(value, name) {
-    const {select, onValueChange} = this.props;
-
-    select[name] = value;
-
-    onValueChange(select, 'select');
-  }
-
-  onTrendChange(value, name) {
-    const {trend, onValueChange} = this.props;
-
-    trend[name] = value;
-
-    onValueChange(trend, 'trend');
-  }
-
-  render() {
-    const {
-      config,
-      families = [],
-      trend,
-      select,
-      onEditConfigFamilyClick,
-    } = this.props;
-
-    return (
-      <Section
-        foldable
-        title={_('Edit Network Vulnerability Test Families ({{counts}})', {
-          counts: families.length,
-        })}
-      >
-        <Table>
-          <TableHeader>
-            <TableRow>
-              <TableHead>{_('Family')}</TableHead>
-              <TableHead>{_('NVTs selected')}</TableHead>
-              <TableHead>{_('Trend')}</TableHead>
-              <TableHead width="9em">{_('Select all NVTs')}</TableHead>
-              <TableHead align="center">{_('Actions')}</TableHead>
-            </TableRow>
-          </TableHeader>
-          <TableBody>
-            {families.map(family => {
-              const {name} = family;
-              return (
-                <NvtFamily
-                  key={name}
-                  config={config}
-                  family={family}
-                  trend={trend[name]}
-                  select={select[name]}
-                  onEditConfigFamilyClick={onEditConfigFamilyClick}
-                  onSelectChange={this.onSelectChange}
-                  onTrendChange={this.onTrendChange}
-                />
-              );
-            })}
-            <TableRow>
-              <TableData>
-                {_('Total: {{count}}', {count: config.families.count})}
-              </TableData>
-              <TableData align="start">
-                {_('{{known}} of {{max}}', config.nvts)}
-              </TableData>
-              {/* add empty cells to spread row to end of table */}
-              <TableData />
-              <TableData />
-              <TableData />
-            </TableRow>
-          </TableBody>
-        </Table>
-      </Section>
-    );
-  }
-}
-
-NvtFamilies.propTypes = {
-  config: PropTypes.model.isRequired,
-  families: PropTypes.array.isRequired,
-  select: PropTypes.object.isRequired,
-  trend: PropTypes.object.isRequired,
-  onEditConfigFamilyClick: PropTypes.func,
-  onValueChange: PropTypes.func,
-};
 
 const createScannerPreferenceValues = (preferences = []) => {
   const values = {};
